MB-62469: Avoid implicit conversions to integer in JSON
Add to_json or static_cast<int> where needed, to avoid the use of
implicit conversions by nlohmann::json (so we can disable those).

Reviewed-on: https://review.couchbase.org/c/kv_engine/+/211795
Tested-by: Build Bot <build@couchbase.com>

+++ engines/ep/src/stored-value.cc
@@ -535,7 +535,7 @@ void to_json(nlohmann::json& json, const StoredValue& sv) {
     std::stringstream ss;
     ss << sv.getKey();
     json["key"] = ss.str();
-    json["committed"] = sv.getCommitted();
+    json["committed"] = static_cast<int>(sv.getCommitted());
 }
 
 static std::string getSystemEventsValueFromStoredValue(const StoredValue& sv) {
